Added new player sprites and updated code
I added the player sprite sheet an spliced it....ton of work (each
sprite has the correct name and size)

I refactored the playerController and updated couple things.  We may
have to change to GroundCheck again if we want to be able to fall
through objects.

Added material "Slick" so our player does not stick to vertical surfaces
when in air and force is applied.