Level 0 Tilemap
Created a sample level using Tiled.  Currently there are no items on the
map, and one of the "puzzles" simple blocks off part of the level until
we can pull it into Unity.  This level should let us test all sorts of
collision, some slopes, and more.
The idea is for the player to make his way right, then up, then left to
get a powerup that let's them break the darker wall in the middle of the
map.  Then the proceed down to get a key, and finally up to the middle
section and right to open a chest.

Added under Assets/Tilemaps but don't know how to actually add it to a
scene just yet.  Need to research and then experiment once I can think
straight.  Too much Nyquil