Level progress!
I have two versions of level 0.  On the left of my scene is the
"original" with a quick and dirty collision layer.  On the right is a
newer version IN PROGRESS where I am mapping collision to each tile
type, so as levels are created in the future, collision does not have to
be drawn on top.  I want to add another copy of each tileset to Tiled
but leave off collision for cases where we don't want collision, or want
to layer it on manually.  Again, the collision on the new version is a
WIP, so be prepared to fall through the level at the moment.